get_reply_viewを関数化, guild_adminのmodal
